Refactor liveshare setup to a new function
